refactor: package by feature 적용 시도.
