+++ src/main/java/nextstep/subway/etc/PageController.java
@@ -1,4 +1,4 @@
-package nextstep.subway.ui;
+package nextstep.subway.etc;
 
 import org.springframework.http.MediaType;
 import org.springframework.stereotype.Controller;
