+++ src/main/java/nextstep/subway/line/LineService.java
@@ -1,20 +1,16 @@
-package nextstep.subway.application;
+package nextstep.subway.line;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.LineRepository;
-import nextstep.subway.domain.Section;
-import nextstep.subway.domain.Sections;
-import nextstep.subway.domain.Station;
-import nextstep.subway.domain.StationRepository;
-import nextstep.subway.dto.LineRequest;
-import nextstep.subway.dto.LineResponse;
-import nextstep.subway.dto.SectionRequest;
-import nextstep.subway.dto.SectionResponse;
-import nextstep.subway.dto.UpdateLineRequest;
-import nextstep.subway.exception.NoLineException;
-import nextstep.subway.exception.NoStationException;
+import nextstep.subway.line.dto.LineRequest;
+import nextstep.subway.line.dto.LineResponse;
+import nextstep.subway.line.dto.SectionRequest;
+import nextstep.subway.line.dto.SectionResponse;
+import nextstep.subway.line.dto.UpdateLineRequest;
+import nextstep.subway.line.exception.NoLineException;
+import nextstep.subway.line.exception.NoStationException;
+import nextstep.subway.station.Station;
+import nextstep.subway.station.StationRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
