+++ src/main/java/nextstep/subway/line/Section.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.line;
 
 import java.util.Arrays;
 import java.util.List;
@@ -8,7 +8,8 @@ import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.ManyToOne;
-import nextstep.subway.exception.IllegalDistanceException;
+import nextstep.subway.line.exception.IllegalDistanceException;
+import nextstep.subway.station.Station;
 import org.hibernate.annotations.Fetch;
 import org.hibernate.annotations.FetchMode;
