+++ src/main/java/nextstep/subway/line/dto/LineRequest.java
@@ -1,9 +1,9 @@
-package nextstep.subway.dto;
+package nextstep.subway.line.dto;
 
-import nextstep.subway.domain.Line;
-import nextstep.subway.domain.Section;
-import nextstep.subway.domain.Sections;
-import nextstep.subway.domain.Station;
+import nextstep.subway.station.Station;
+import nextstep.subway.line.Line;
+import nextstep.subway.line.Section;
+import nextstep.subway.line.Sections;
 
 public class LineRequest {
