+++ src/main/java/nextstep/subway/line/exception/IllegalDistanceException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.exception;
 
 public class IllegalDistanceException extends IllegalArgumentException {
