+++ src/main/java/nextstep/subway/line/exception/SameStationException.java
@@ -1,4 +1,4 @@
-package nextstep.subway.exception;
+package nextstep.subway.line.exception;
 
 public class SameStationException extends IllegalArgumentException {
