+++ src/main/java/nextstep/subway/station/Station.java
@@ -1,4 +1,4 @@
-package nextstep.subway.domain;
+package nextstep.subway.station;
 
 import java.util.Objects;
 import javax.persistence.Column;
@@ -6,6 +6,7 @@ import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import nextstep.subway.etc.BaseEntity;
 
 @Entity
 public class Station extends BaseEntity {
