+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -8,8 +8,8 @@ import static org.assertj.core.api.Assertions.assertThat;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
 import nextstep.subway.AcceptanceTest;
-import nextstep.subway.dto.SectionResponse;
-import nextstep.subway.dto.StationResponse;
+import nextstep.subway.line.dto.SectionResponse;
+import nextstep.subway.station.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
